feat : Stomp Board 통신 및 에러 해결

+++ src/server/src/main/java/com/econovation/recruit/application/port/in/BoardUseCase.java
@@ -12,7 +12,7 @@ public interface BoardUseCase {
 
     Map<String, Integer> getNewestLocationByNavLocAndColLoc(Integer navLoc, Integer colLoc);
 
-    Board findById(Long id);
+    Board findById(Integer id);
 
     boolean isDuplicateLocation(Integer navLoc, Integer colLoc, Integer lowLoc);
